feat(run): add Expected column and test comments, drop Objective/Notes
Rework the run results table to align with the progress view:
- add the Expected column, render test comments via a dedicated
test-comments component, persist column-visibility preferences,
and remove the Objective and Notes columns.
Deep links from the progress matrix land here pre-filtered by result
column.
