+++ libs/bublik/features/run/src/lib/run-table/columns/badge-columns.tsx
@@ -2,7 +2,11 @@
 /* SPDX-FileCopyrightText: 2021-2023 OKTET Labs Ltd. */
 import { RunTableColumnConfig, ColumnId } from '../types';
 
-import { RESULT_PROPERTIES, RESULT_TYPE } from '@/shared/types';
+import {
+	RESULT_PROPERTIES,
+	RESULT_TYPE,
+	ResultTableFilter
+} from '@/shared/types';
 import { BadgeVariants, Icon } from '@/shared/tailwind-ui';
 import {
 	getAbnormal,
@@ -207,3 +211,14 @@ const columnConfigs: RunTableColumnConfig[] = [
 ];
 
 export const badgeColumns = columnConfigs.map(createRunColumn);
+
+export function getResultFilterForColumnId(
+	columnId: ColumnId | string
+): ResultTableFilter {
+	const config = columnConfigs.find((column) => column.id === columnId);
+
+	return {
+		results: config?.results ?? [],
+		resultProperties: config?.resultProperties ?? []
+	};
+}
